Bunch of fixes and updates Fix exception when guild didn't have any async config Add seed hash output image to seed gen Add 64 bit seed support Remove direct references from project, opting for DLL references instead

+++ MinishCapRandomizerSeedGeneratorBot/Threading/Models/SeedGenerationRequest.cs
@@ -5,7 +5,7 @@ namespace MinishCapRandomizerSeedGeneratorBot.Threading.Models;
 public class SeedGenerationRequest
 {
     public SocketSlashCommand Command { get; set; }
-    public int Seed { get; set; }
+    public ulong Seed { get; set; }
     public string SettingsString { get; set; }
     public string CosmeticsString { get; set; }
     public bool UploadSpoiler { get; set; }
